requirements.txt
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-experimental>=0.0.20
streamlit>=1.28.0
